pyproject.toml
[tool.poetry]
name = "bet-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.1.5"
requests = "^2.27.1"
python-decouple = "^3.6"


[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
